feat(frontends/lean/elaborator): elaborate ```(e) to type `reflected e` if possible and add coercion reflected -> expr
