+++ src/frontends/lean/builtin_exprs.cpp
@@ -651,7 +651,7 @@ static expr parse_lazy_quoted_pexpr(parser_state & p, unsigned, expr const *, po
         e = mk_typed_expr_distrib_choice(p, t, e, pos);
     }
     p.check_token_next(get_rparen_tk(), "invalid quoted expression, `)` expected");
-    return p.save_pos(mk_pexpr_quote(e), pos);
+    return p.save_pos(mk_quote(e, /* is_expr */ false), pos);
 }
 
 static expr parse_quoted_pexpr(parser_state & p, unsigned, expr const *, pos_info const & pos) {
@@ -665,7 +665,7 @@ static expr parse_quoted_pexpr(parser_state & p, unsigned, expr const *, pos_inf
         e = mk_typed_expr_distrib_choice(p, t, e, pos);
     }
     p.check_token_next(get_rparen_tk(), "invalid quoted expression, `)` expected");
-    return p.save_pos(mk_pexpr_quote(e), pos);
+    return p.save_pos(mk_quote(e, /* is_expr */ false), pos);
 }
 
 static expr parse_quoted_expr(parser_state & p, unsigned, expr const *, pos_info const & pos) {
@@ -682,11 +682,8 @@ static expr parse_quoted_expr(parser_state & p, unsigned, expr const *, pos_info
         }
         p.check_token_next(get_rparen_tk(), "invalid quoted expression, `)` expected");
     }
-    e = mk_quote_core(e, true);
-    if (!p.in_pattern()) {
-        e = elaborate_quote(e, p.env(), p.get_options(), /* in_pattern */ false);
-        // note: when `p.in_pattern()`, quote will be elaborated in `parser::patexpr_to_{pattern,expr}`
-    }
+    // do not introduce `expr.subst` calls in patterns
+    e = p.in_pattern() ? mk_quote_core(e, true) : mk_quote(e, true);
     return p.save_pos(e, pos);
 }
