+++ src/library/quote.h
@@ -8,10 +8,11 @@ Author: Leonardo de Moura
 #include "kernel/expr.h"
 
 namespace lean {
-expr mk_pexpr_quote(expr const &e);
+expr mk_quote(expr const & e, bool is_expr);
 bool is_quote(expr const & e);
 bool is_expr_quote(expr const &e);
 expr const & get_quote_expr(expr const & e);
+expr const & get_quote_type(expr const & e);
 expr mk_quote_core(expr const & e, bool is_expr);
 expr mk_reflected(expr const & e, expr const & ty, level const & l);
