+++ src/library/vm/vm_pexpr.cpp
@@ -50,7 +50,7 @@ vm_obj pexpr_pos(vm_obj const & e) {
 }
 
 vm_obj pexpr_mk_quote_macro(vm_obj const & e) {
-    return to_obj(mk_pexpr_quote(to_expr(e)));
+    return to_obj(mk_quote(to_expr(e), /* is_expr */ false));
 }
 
 vm_obj pexpr_mk_prenum_macro(vm_obj const & n) {
