+++ tests/lean/eval_expr_error.lean.expected.out
@@ -1,8 +1,4 @@
-eval_expr_error.lean:5:8: error: failed to synthesize type class instance for
-A : Type,
-tst1 : tactic unit,
-a : expr
-⊢ reflected A
+eval_expr_error.lean:5:18: error: unexpected local in quotation expression
 eval_expr_error.lean:8:0: error: invalid eval_expr, type mismatch
 state:
 ⊢ true
